+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 - Support for Wazuh 5.0.0
 - Added debounce to Document Level Query text inputs
-- Add "Active Response" monitor type [#33](https://github.com/wazuh/wazuh-dashboard-alerting/pull/33)
+- Add "Active Response" monitor type [#33](https://github.com/wazuh/wazuh-dashboard-alerting/pull/33) [#133](https://github.com/wazuh/wazuh-dashboard-alerting/issues/133) 
 
 ### Changed